+++ packages/expo-dev-menu/vendored/react-native-reanimated/Common/cpp/headers/Tools/DevMenuJSIStoreValueUser.h
@@ -6,7 +6,7 @@
 #include <mutex>
 #include <unordered_map>
 #include <vector>
-#include "Scheduler.h"
+#include "DevMenuScheduler.h"
 
 using namespace facebook;