+++ docs/components/plugins/api/APISectionProps.tsx
@@ -11,7 +11,7 @@ import {
   TypeDefinitionData,
 } from '~/components/plugins/api/APIDataTypes';
 import { APISectionDeprecationNote } from '~/components/plugins/api/APISectionDeprecationNote';
-import { PlatformTags } from '~/components/plugins/api/APISectionPlatformTags';
+import { APISectionPlatformTags } from '~/components/plugins/api/APISectionPlatformTags';
 import {
   CommentTextBlock,
   getCommentOrSignatureComment,
@@ -105,7 +105,7 @@ export const renderProp = (
   return (
     <div key={`prop-entry-${name}`} css={[STYLES_APIBOX, !exposeInSidebar && STYLES_APIBOX_NESTED]}>
       <APISectionDeprecationNote comment={extractedComment} />
-      <PlatformTags comment={comment} prefix="Only for:" firstElement />
+      <APISectionPlatformTags comment={comment} prefix="Only for:" firstElement />
       <HeaderComponent>
         <InlineCode customCss={!exposeInSidebar ? STYLES_NOT_EXPOSED_HEADER : undefined}>
           {name}
@@ -130,7 +130,7 @@ const APISectionProps = ({ data, defaultProps, header = 'Props' }: APISectionPro
   const baseProp = data.find(prop => prop.name === header);
   return data?.length ? (
     <>
-      {header === 'Props' ? (
+      {data?.length === 1 || header === 'Props' ? (
         <H2 key="props-header">{header}</H2>
       ) : (
         <div>
@@ -142,7 +142,7 @@ const APISectionProps = ({ data, defaultProps, header = 'Props' }: APISectionPro
         </div>
       )}
       {data.map((propsDefinition: PropsDefinitionData) =>
-        renderProps(propsDefinition, defaultProps, header === 'Props')
+        renderProps(propsDefinition, defaultProps, data?.length === 1 || header === 'Props')
       )}
     </>
   ) : null;