+++ packages/expo-dev-menu/vendored/react-native-reanimated/Common/cpp/Tools/WorkletEventHandler.cpp
@@ -1,4 +1,4 @@
-#include "WorkletEventHandler.h"
+#include "DevMenuWorkletEventHandler.h"
 
 namespace devmenureanimated {