+++ packages/expo-dev-menu/vendored/react-native-reanimated/ios/native/DevMenuREAIOSScheduler.mm
@@ -1,5 +1,5 @@
 #import "DevMenuREAIOSScheduler.h"
-#import "RuntimeManager.h"
+#import "DevMenuRuntimeManager.h"
 
 namespace devmenureanimated {