+++ packages/expo-dev-menu/CHANGELOG.md
@@ -10,6 +10,12 @@
 
 ### 💡 Others
 
+## 1.1.1 — 2022-07-20
+
+### 🐛 Bug fixes
+
+- Fixed compatibility with the `react-native-reanimated` on iOS. ([#18306](https://github.com/expo/expo/pull/18306) by [@lukmccall](https://github.com/lukmccall))
+
 ## 1.1.0 — 2022-07-18
 
 ### 🎉 New features